Split out testlib
