+++ src/driver/CMakeLists.txt
@@ -38,13 +38,12 @@ install(TARGETS  genny
 # Define our test for the driver
 add_executable(test_driver
     test/Driver_test.cpp
-    $<TARGET_OBJECTS:catch2>
 )
 target_include_directories(test_driver
     PRIVATE
         src
 )
-target_link_libraries(test_driver driver gennylib_testutil)
+target_link_libraries(test_driver driver testlib)
 
 ParseAndAddCatchTests(test_driver)
