+++ src/gennylib/test/RateLimiter_benchmark.cpp
@@ -1,6 +1,16 @@
-#include "test.h"
-
-#include "ActorHelper.hpp"
+// Copyright 2019-present MongoDB Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #include <iostream>
 
@@ -11,6 +21,8 @@
 #include <gennylib/config/RateLimiterOptions.hpp>
 #include <gennylib/conventions.hpp>
 #include <gennylib/v1/RateLimiter.hpp>
+#include <testlib/ActorHelper.hpp>
+#include <testlib/helpers.hpp>
 
 namespace Catchers = Catch::Matchers;
