+++ src/gennylib/test/orchestrator_test.cpp
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "test.h"
-
 #include <chrono>
 #include <iostream>
 #include <mutex>
@@ -27,6 +25,7 @@
 #include <gennylib/Orchestrator.hpp>
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
+#include <testlib/helpers.hpp>
 
 using namespace genny;
 using namespace std::chrono;
