--- src/gennylib/test/test.h
@@ -1,1 +1,0 @@
-#include <helpers.hpp>
