+++ src/metrics/CMakeLists.txt
@@ -36,10 +36,9 @@ install(TARGETS  metrics
 
 add_executable(test_metrics
     test/metrics_test.cpp
-    $<TARGET_OBJECTS:catch2>
 )
 
-target_link_libraries(test_metrics metrics gennylib_testutil)
+target_link_libraries(test_metrics metrics catch2 testlib)
 
 ParseAndAddCatchTests(test_metrics)
