+++ src/testlib/src/ActorHelper.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ActorHelper.hpp"
+#include <testlib/ActorHelper.hpp>
 
 #include <thread>
 #include <vector>
